requirements.txt
Django==2.1.4
django-appconf==1.0.3
django-ckeditor==5.7.1
django-crispy-forms==1.8.1
django-imagekit==4.0.2
django-js-asset==1.2.2
django-widget-tweaks==1.4.3
ll==1.0
pilkit==2.0
Pillow==5.3.0
pkg-resources==0.0.0
python-slugify==4.0.0
pytz==2018.7
six==1.14.0
slugify==0.0.1
text-unidecode==1.3
